Move unit test classes from it-pipeline to core module
